Assets allows parameters, chunks are now wrapped in a div, modified admin mode
Assets do now allow to add further attributes to the tags (for example
for require.js)- Chunks are now wrapped with a div tag which has an id
of scope_name, which is unique. This is needed for positioning the
config button and so on. Added basic positioning of config-buttons.