Chunks can serve own views, new openForm()-Method, Search "improved"
The chunks folder is now added to the view config file.
Further a new openForm()-Method was added to the abstract Chunk-class.
This allows to create the opening form tag. Especially useful if you do
not want to use the form package.
Also added foundation for searchresults chunk. Searching is more
complex than somone would imagine. It is really the show case for all
the awesome goodies of this CMS.