Made the chunk system lazy
In order to provide efficient caching, it is important, that we reduce
database calls. Therefore the Chunk class was modified to only load the
content field of a chunk if it is called. Further added a
ChunkNotFoundException which is fired when you access the content
property and the chunk does not exist.