Rewrote menu chunk to use nested sets, dropped forms from contacts, 404
Several changes indeed :)
First: the menu chunk is now completely rewritten and is now using a
nested set to store the menu structure. This improves the performance
tremendously and gives us a lot more flexibilty! Further the contact
chunk is now able to use views for the contact form. It's time to drop
the forms lib though...
The text chunk now caches only for 10 minutes.
I've tried to get a nicer 404-message but it's in fact not working
right now :(
The openForm-Chunk-method got a new property $attributes which is an
array where you can specify further attributes of the form-tag like
class name etc.