Updated the application to the current versions
As this piece of software is using software, which is still in
development, there sometimes have to be some changes in the API. That
was the case with sentry. I also took the chance and renamed the Sentry
alias to Authentication. That will make it much easier for
Chunk-Developers to work with this module.