Complete rewrote of the chunk system
The chunk system is now way more flexible. Instead of only calling a
function, which creates an instance of the class and calls the show
method, there is now a ChunkManager, which does all the loading and
really knows which modules are loaded. That was really the biggest
issue of the previous system.
Now there are also things possible like "inter-chunk-communication" and
non-visual chunks (statistics chunk, e.g.).
For this new chunk management, the Container chunk had to be nearly
completely rewritten.
Further dropped the raftalks/Form class. Which was not really flexible.