The Frontend is now completely converted to Twig
Yes, seriously! I think it's a better choice to integrate a more
powerful template engine, especially as Blade and Twig can coexist so
nicely! So it's up to the user, which engine to use although we do
favour Twig over Blade (especially due to the strange escaping issue of
Blade...)