openForm is now in ChunkManager, Login Chunk for User Login
The openForm-method, which was placed in the abstract Chunk class, now
moved into the ChunkManager. But we have declared a twig alias
open_form which makes it possible to define the form options in your
view file. Isn't that neat? :D All chunks, which had a form, are now
using this new method.
There is also a new Login Chunk, which you can use to create a frontend
login for a member area or something like that. Pretty neat, isn't it?
:D