First steps towards cool media management!
The first steps are made on the journey to a cool media management.
There is now a whole library "Media" which contains several classes.
Entry point is the MediaAccessor. Right now it can only lookup a file
in the virtual file system. All virtual pathes have the format:
driver/path/to/the/file. So it extracts the driver from the first
component and looks, if there is a registered driver for this key.
Each driver has to implement the Cmex\Media\DriverInterface. If it's
also listed in app/config/mediadrivers.php it will be automatically on
startup. Therefore we are using a service provider, which instantiates
the MediaAccessor for us, reads all the drivers and adds the class to
the IOC-Container. By using a Laravel Facade, the class is then
available via the Media Facade.