Fixed a design flaw and removed senseless method
The drivers can have config items, but these should not be read with
Config::get() in the constructor, because it creates an unneeded
dependency for the drivers.
Instead the config is injected in the MediaServiceProvider.
Further the getFileForId-Method was removed, as it didn't make sense to
have another way of referencing the file. The virtual file tree should
be enough therefore.