Urgent bux fix in ChunkManager and made TwigConnector more secure
When moving the openForm-Method into the ChunkManager, there had to be
a way, that the ChunkManager knows, which Chunk is being executed.
Therefore there was a property currentChunk. Actually I forgot, that
there is also a container widget, so that the chunk manager has to
remember, that this chunk is still executed after a subchunk finished.
So now we are using an execution stack to solve this problem. The
container chunk also had to be modified, as there was the wrong method
called...
The TwigConnector was improved a little bit to only load functions,
that are "registered".
There are also some extra experiments for the media manager included.