Chunks are now autoloaded without composer
For now the chunks are now autoloaded by the Laravel ClassLoader.
That's not perfect as it is performing a file_exists on every class it
loads. So it would be better to create an intelligent ChunkLoader,
which fetches a list of installed chunks (and can expand it on it's
own), but therefore we first of all have to finish the chunk structure
etc. So there is now no "composer dumpautoload" needed anymore for
adding a new chunk. Yay! :)