[IMPORTANT] Changed structure
Reorganisation of chunks and libraries, which are now stored in the
modules-folder.
This commit needs a composer dumpautoload AND php artisan db:seed, as
the Chunks where reorganised into packages (i.e. Text chunk is now
Text.Block or Container is now Container.Standard) and the database
entries had to be refreshed.