Added artisan module commands and improved ModuleServiceProvider
Now the ModuleServiceProvider is only adding the modules listed in
app/storage/modules.json. If this file is not present, it will try to
create it.
There are now some artisan commands for creating a new module (php
artisan module:make), adding an existing module directory (php artisan
module:add), which also adds the needed files if not present and for
refreshing the app/storage/modules.json.