Login now uses Laravel's FormBuilder, removed Sentry-Migrations
If you login with wrong credentials, these are filled in the form again
for you.
The Sentry migrations are now not in the database/migrations folder
anymore because they were always colliding with updated ones. You will
now also have to migrate these on your own (for now), see readme.md for
details.